backend/requirements.txt
aioquic==1.2.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.10.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==5.0.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.8.0
eventlet==0.40.4
executing==2.1.0
fastjsonschema==2.20.0
Flask==3.1.2
Flask-SocketIO==5.5.1
fonttools==4.53.1
fqdn==1.5.1
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.1.0
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaitaistruct==0.11
kiwisolver==1.4.7
ldap3==2.9.1
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mitmproxy==12.2.0
mitmproxy-windows==0.12.7
mitmproxy_rs==0.12.7
msgpack==1.1.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.1
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
publicsuffix2==2.20191221
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydivert==2.1.0
Pygments==2.18.0
pylsqpack==0.3.23
pyOpenSSL==25.3.0
pyparsing==3.1.4
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-json-logger==2.0.7
python-socketio==5.15.0
pytz==2024.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.15
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==74.1.2
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tornado==6.5.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.14.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
urwid==3.0.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wsproto==1.2.0
xgboost==3.0.4
zstandard==0.25.0
flask
flask-socketio
python-socketio
flask-jwt-extended
flask-bcrypt
flask-cors
python-dotenv
sqlalchemy
